Move password-reset service methods to own service
